clean up

+++ server-express-mysql/models/user.js
@@ -12,7 +12,7 @@ module.exports = (sequelize, DataTypes) => {
       lName: DataTypes.STRING,
       email: DataTypes.STRING,
       password: DataTypes.STRING,
-      phone: DataTypes.INTEGER
+      phone: DataTypes.BIGINT
 
     },
     {}
